backend/requirements.txt
# Python 3.11.8 / 3.13 compatible versions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
PyMySQL==1.1.0
PyJWT==2.8.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
mysqlclient==2.2.4  # More stable than PyMySQL